feat: add rendercv python module
This commit initializes the python module `rendercv` at
version 1.14. It employs the dream2nix pip module to fetch
binary wheels from PyPi. A top-level alias for `rendercv`
is made for ease of access.

How packages are arranged and evaluated was also reworked.
We now manually specific the package paths and call `evalModule`
instead of using the `importPackages` function. It allows for more
flexibility in have non-flat directory structures and scoped package
sets. The accumulated package set is finally flattened into a single
level attribute set compatible with the flake schma.