fix: fix canon-cups-ufr2 build failure due to broken symlinks
The `noBrokenSymlinks` setup hook introduced recently to nixpkgs cause
the canon-cups-ufr2 package to fail to build. This is because the cngplp
component is never built but is referenced by installed files in the output
via symlinks. This commit ports the PR by lluchs which refactors the build
process to reuse the bundled RPM spec file and build cngplp correctly, which
eliminates the dangling symlinks in the output.